gerber file for the pcb shield
the shield should be able to handle both SPI and LPC sniffing. For SPI sniffing the header socket or the pin rows can be used to connect a test clip. For the original LPC sniffing the pogo pins as in the original design can be used.
The dip switch is used to select between SPI and LPC sniffing.
The push button is for convinience to make switching into boot mode for reprogramming a bit easier.